#926 Necessary corrections without which the database migration does not work. The test environment on the docker did not work without migration. Improved method invocation in UserDwr, Mutes the interface name IUserDAO to UserDAO. Adding uniqe to the user's name in the database and commenting on the code that adds listwatch permissions (which did not work during user editing).
