+++ src/org/scada_lts/dao/UserDAO.java
@@ -1,290 +1,65 @@
+/*
+ * (c) 2019 VROC.ai https://vroc.ai/
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package org.scada_lts.dao;
 
-
 import com.serotonin.mango.vo.User;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.ArgumentPreparedStatementSetter;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import com.mysql.jdbc.Statement;
-
 import java.util.List;
 
 /**
- * DAO for User.
- *
- * @author grzegorz bylica Abil'I.T. development team, sdt@abilit.eu
- *         person supporting and coreecting translation Jerzy Piejko
- * @author Mateusz Kaproń Abil'I.T. development team, sdt@abilit.eu
+ * @autor Khelifi Hassene,
+ * @autor grzegorz.bylica@gmail.com (SoftQ) on 03.07.19
  */
-public class UserDAO {
+public interface UserDAO {
+    List<Integer> getAll();
 
-	private static final Log LOG = LogFactory.getLog(UserDAO.class);
+    User getUser(int id);
 
-	protected final static String COLUMN_NAME_ID = "id";
-	protected final static String COLUMN_NAME_USERNAME = "username";
-	protected final static String COLUMN_NAME_PASSWORD = "password";
-	protected final static String COLUMN_NAME_EMAIL = "email";
-	protected final static String COLUMN_NAME_PHONE = "phone";
-	protected final static String COLUMN_NAME_ADMIN = "admin";
-	protected final static String COLUMN_NAME_DISABLED = "disabled";
-	protected final static String COLUMN_NAME_SELECTED_WATCH_LIST = "selectedWatchList";
-	protected final static String COLUMN_NAME_HOME_URL = "homeUrl";
-	protected final static String COLUMN_NAME_LAST_LOGIN = "lastLogin";
-	protected final static String COLUMN_NAME_RECEIVE_ALARM_EMAILS = "receiveAlarmEmails";
-	protected final static String COLUMN_NAME_RECEIVE_OWN_AUDIT_EVENTS = "receiveOwnAuditEvents";
+    User getUser(String username);
 
-	// @formatter:off
-	protected static final String USER_SELECT_ID = ""
-			+ "select "
-			+ COLUMN_NAME_ID + " "
-			+ "from users ";
+    List<User> getUsers();
 
-	protected static final String USER_SELECT = ""
-			+ "select "
-				+ COLUMN_NAME_ID + ", "
-				+ COLUMN_NAME_USERNAME + ", "
-				+ COLUMN_NAME_PASSWORD + ", "
-				+ COLUMN_NAME_EMAIL + ", "
-				+ COLUMN_NAME_PHONE + ", "
-				+ COLUMN_NAME_ADMIN + ", "
-				+ COLUMN_NAME_DISABLED + ", "
-				+ COLUMN_NAME_SELECTED_WATCH_LIST + ", "
-				+ COLUMN_NAME_HOME_URL + ", "
-				+ COLUMN_NAME_LAST_LOGIN + ", "
-				+ COLUMN_NAME_RECEIVE_ALARM_EMAILS + ", "
-				+ COLUMN_NAME_RECEIVE_OWN_AUDIT_EVENTS + " "
-			+ "from users ";
+    List<User> getActiveUsers();
 
-	protected static final String USER_SELECT_ORDER = ""
-			+ USER_SELECT
-			+ "order by username ";
+    void updateHomeUrl(int userId, String homeUrl);
 
-	protected static final String USER_SELECT_WHERE_ID = ""
-			+ USER_SELECT
-			+ "where "
-				+ COLUMN_NAME_ID + "=? ";
+    void updateLogin(int userId);
 
-	protected static final String USER_SELECT_WHERE_USERNAME = ""
-			+ USER_SELECT
-			+ "where lower("
-				+ COLUMN_NAME_USERNAME + ")=?";
+    String getUserTimezone(int id);
 
-	protected static final String USER_SELECT_ACTIVE = ""
-			+ USER_SELECT
-			+ "where "
-				+ COLUMN_NAME_DISABLED + "=? ";
+    String getUserZone(int id);                                                                                 ///smart e-tech //time_zone
 
-	protected static final String USER_INSERT = ""
-			+ "insert into users ("
-				+ COLUMN_NAME_USERNAME + ", "
-				+ COLUMN_NAME_PASSWORD + ", "
-				+ COLUMN_NAME_EMAIL + ", "
-				+ COLUMN_NAME_PHONE + ", "
-				+ COLUMN_NAME_ADMIN + ", "
-				+ COLUMN_NAME_DISABLED + ", "
-				+ COLUMN_NAME_HOME_URL + ", "
-				+ COLUMN_NAME_RECEIVE_ALARM_EMAILS + ", "
-				+ COLUMN_NAME_RECEIVE_OWN_AUDIT_EVENTS + ") "
-			+ "values (?,?,?,?,?,?,?,?,?) ";
+    User getUserByMail(String email);
 
-	protected static final String USER_UPDATE = ""
-			+ "update users set "
-				+ COLUMN_NAME_USERNAME + "=?, "
-				+ COLUMN_NAME_PASSWORD + "=?, "
-				+ COLUMN_NAME_EMAIL + "=?, "
-				+ COLUMN_NAME_PHONE + "=?, "
-				+ COLUMN_NAME_ADMIN + "=?, "
-				+ COLUMN_NAME_DISABLED + "=?, "
-				+ COLUMN_NAME_HOME_URL + "=?, "
-				+ COLUMN_NAME_RECEIVE_ALARM_EMAILS + "=?, "
-				+ COLUMN_NAME_RECEIVE_OWN_AUDIT_EVENTS + "=? "
-			+ "where "
-				+ COLUMN_NAME_ID + "=? ";
+    void updateUserTimezone(int userId, String timezone);
 
-	protected static final String USER_UPDATE_LOGIN = ""
-			+ "update users set "
-				+ COLUMN_NAME_LAST_LOGIN + "=? "
-			+ "where "
-				+ COLUMN_NAME_ID + "=? ";
+    void updateUserZone(int userId, String zone);
 
-	protected static final String USER_UPDATE_HOME_URL = ""
-			+ "update users set "
-				+ COLUMN_NAME_HOME_URL + "=? "
-			+ "where "
-				+ COLUMN_NAME_ID + "=? ";
+    @Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
+    int insert(User user);
 
-	protected static final String USER_DELETE = ""
-			+ "delete from users where "
-			+ COLUMN_NAME_ID + "=? ";
-	// @formatter:on
+    @Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
+    void update(User user);
 
-	protected class UserRowMapper implements RowMapper<User> {
-
-		@Override
-		public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-			User user = new User();
-			user.setId(rs.getInt(COLUMN_NAME_ID));
-			user.setUsername(rs.getString(COLUMN_NAME_USERNAME));
-			user.setPassword(rs.getString(COLUMN_NAME_PASSWORD));
-			user.setEmail(rs.getString(COLUMN_NAME_EMAIL));
-			user.setPhone(rs.getString(COLUMN_NAME_PHONE));
-			user.setAdmin(DAO.charToBool(rs.getString(COLUMN_NAME_ADMIN)));
-			user.setDisabled(DAO.charToBool(rs.getString(COLUMN_NAME_DISABLED)));
-			user.setSelectedWatchList(rs.getInt(COLUMN_NAME_SELECTED_WATCH_LIST));
-			user.setHomeUrl(rs.getString(COLUMN_NAME_HOME_URL));
-			user.setLastLogin(rs.getLong(COLUMN_NAME_LAST_LOGIN));
-			user.setReceiveAlarmEmails(rs.getInt(COLUMN_NAME_RECEIVE_ALARM_EMAILS));
-			user.setReceiveOwnAuditEvents(DAO.charToBool(rs.getString(COLUMN_NAME_RECEIVE_OWN_AUDIT_EVENTS)));
-			return user;
-		}
-	}
-
-	public List<Integer> getAll() {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("getAll()");
-		}
-
-		return DAO.getInstance().getJdbcTemp().queryForList(USER_SELECT_ID, Integer.class);
-	}
-
-	public User getUser(int id) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("getUser(int id) id:" + id);
-		}
-
-		User user;
-		try {
-			user = DAO.getInstance().getJdbcTemp().queryForObject(USER_SELECT_WHERE_ID, new Object[]{id}, new UserRowMapper());
-		} catch (EmptyResultDataAccessException e) {
-			user = null;
-		}
-		return user;
-	}
-
-	public User getUser(String username) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("getUser(String username) username:" + username);
-		}
-
-		User user;
-		try {
-			user = DAO.getInstance().getJdbcTemp().queryForObject(USER_SELECT_WHERE_USERNAME, new Object[]{username}, new UserRowMapper());
-		} catch (EmptyResultDataAccessException e) {
-			user = null;
-		}
-		return user;
-	}
-
-	public List<User> getUsers() {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("getUsers()");
-		}
-
-		return DAO.getInstance().getJdbcTemp().query(USER_SELECT_ORDER, new UserRowMapper());
-	}
-
-	public List<User> getActiveUsers() {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("getActiveUsers()");
-		}
-
-		return DAO.getInstance().getJdbcTemp().query(USER_SELECT_ACTIVE, new Object[]{DAO.boolToChar(false)}, new UserRowMapper());
-	}
-
-	public void updateHomeUrl(int userId, String homeUrl) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("updateHomeUrl(int userId, String homeUrl) userId:" + userId + ", homeUrl:" + homeUrl);
-		}
-
-		DAO.getInstance().getJdbcTemp().update(USER_UPDATE_HOME_URL, new Object[]{homeUrl, userId});
-	}
-
-	public void updateLogin(int userId) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("updateLogin(int userId) userId:" + userId);
-		}
-
-		DAO.getInstance().getJdbcTemp().update(USER_UPDATE_LOGIN, new Object[]{System.currentTimeMillis(), userId});
-	}
-
-	@Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
-	public int insert(final User user) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("insert(User user) user:" + user.toString());
-		}
-
-		KeyHolder keyHolder = new GeneratedKeyHolder();
-
-		DAO.getInstance().getJdbcTemp().update(new PreparedStatementCreator() {
-			@Override
-			public PreparedStatement createPreparedStatement(Connection connection) throws SQLException {
-				PreparedStatement preparedStatement = connection.prepareStatement(USER_INSERT, Statement.RETURN_GENERATED_KEYS);
-				new ArgumentPreparedStatementSetter(new Object[]{
-						user.getUsername(),
-						user.getPassword(),
-						user.getEmail(),
-						user.getPhone(),
-						DAO.boolToChar(user.isAdmin()),
-						DAO.boolToChar(user.isDisabled()),
-						user.getHomeUrl(),
-						user.getReceiveAlarmEmails(),
-						DAO.boolToChar(user.isReceiveOwnAuditEvents())
-				}).setValues(preparedStatement);
-				return preparedStatement;
-			}
-		}, keyHolder);
-		return keyHolder.getKey().intValue();
-	}
-
-	@Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
-	public void update(final User user) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("update(User user) user:" + user);
-		}
-
-		DAO.getInstance().getJdbcTemp().update(USER_UPDATE, new Object[]{
-				user.getUsername(),
-				user.getPassword(),
-				user.getEmail(),
-				user.getPhone(),
-				DAO.boolToChar(user.isAdmin()),
-				DAO.boolToChar(user.isDisabled()),
-				user.getHomeUrl(),
-				user.getReceiveAlarmEmails(),
-				DAO.boolToChar(user.isReceiveOwnAuditEvents()),
-				user.getId()
-		});
-	}
-
-	@Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
-	public void delete(int userId) {
-
-		if (LOG.isTraceEnabled()) {
-			LOG.trace("delete(int userId) userId:" + userId);
-		}
-
-		DAO.getInstance().getJdbcTemp().update(USER_DELETE, new Object[]{userId});
-	}
+    @Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = SQLException.class)
+    void delete(int userId);
 }
