+++ test/org/scada_lts/service/ViewHierarchyServiceMoveFolderTest.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 import org.scada_lts.dao.TestDAO;
-import org.scada_lts.dao.UserDAO;
+import org.scada_lts.dao.UserDAOImp;
 import org.scada_lts.dao.ViewDAO;
 import org.scada_lts.dao.ViewHierarchyDAO;
 import org.scada_lts.dao.model.viewshierarchy.ViewHierarchyNode;
@@ -98,7 +98,7 @@ public class ViewHierarchyServiceMoveFolderTest extends TestDAO {
 	    assertTrue(vhs.getNode(3).size()==0);
 	    assertTrue(vhs.getNode(ViewHierarchyDAO.ROOT_ID).size()==1);
 	    
-	    UserDAO uDAO = new UserDAO();
+	    UserDAOImp uDAO = new UserDAOImp();
 		
 		User user = new User();
 		user.setAdmin(true);
